feat: Generate unique blog posts to avoid repetition
The code changes in this commit modify the `postsGenerator` function to generate blog posts that do not repeat topics that have already been generated. The function now takes an array of previous posts titles and authors as an argument to ensure uniqueness. The generated content includes a list of already generated posts to avoid duplication.